Rename folders to reflect correct name of chatbot
