+++ build.gradle
@@ -28,11 +28,11 @@ test {
 }
 
 application {
-    mainClass.set("duke.Duke")
+    mainClass.set("miles.Miles")
 }
 
 shadowJar {
-    archiveBaseName = "duke"
+    archiveBaseName = "miles"
     archiveClassifier = null
     dependsOn("distZip", "distTar")
 }
