+++ src/main/java/miles/task/Deadline.java
@@ -1,4 +1,4 @@
-package duke.task;
+package miles.task;
 
 import java.time.LocalDateTime;
