+++ src/test/java/miles/ParserTest.java
@@ -1,4 +1,4 @@
-package duke;
+package miles;
 
 import org.junit.jupiter.api.Test;
